# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JobSearch Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
  success: boolean;
}

export interface SearchQuery {
  query: string;
  category?: string;
}

export type ProfileWorkType =
  (typeof ProfileWorkType)[keyof typeof ProfileWorkType];

export const ProfileWorkType = {
  remote: "remote",
  hybrid: "hybrid",
  both: "both",
} as const;

export interface Profile {
  name: string;
  salary_min: number;
  work_type: ProfileWorkType;
  sources: string[];
  queries: SearchQuery[];
  resume_ref?: string;
}

export type CreateProfileRequestWorkType =
  (typeof CreateProfileRequestWorkType)[keyof typeof CreateProfileRequestWorkType];

export const CreateProfileRequestWorkType = {
  remote: "remote",
  hybrid: "hybrid",
  both: "both",
} as const;

export interface CreateProfileRequest {
  name: string;
  salary_min?: number;
  work_type?: CreateProfileRequestWorkType;
  sources?: string[];
  queries?: SearchQuery[];
}

export interface ProfileListResponse {
  profiles: string[];
  active?: string | null;
}

export interface ActiveProfileResponse {
  active: string | null;
}

export interface SetActiveProfileRequest {
  name: string | null;
}

export interface RunSearchRequest {
  profile_name: string;
  mock?: boolean;
}

export interface SearchStartedResponse {
  job_id: string;
  message: string;
}

export interface JobListing {
  title: string;
  company: string;
  location: string;
  salary?: string;
  description?: string;
  url?: string;
  source: string;
  date_found?: string;
  query?: string;
  work_type?: string;
}

export type SearchResultsResponseStatus =
  (typeof SearchResultsResponseStatus)[keyof typeof SearchResultsResponseStatus];

export const SearchResultsResponseStatus = {
  running: "running",
  completed: "completed",
  error: "error",
} as const;

export interface SearchResultsResponse {
  job_id: string;
  status: SearchResultsResponseStatus;
  jobs: JobListing[];
  total: number;
  filepath?: string;
  error?: string;
}

export interface ResumeParseResponse {
  titles: string[];
  skills: string[];
  tools: string[];
  certifications: string[];
  industries: string[];
  suggested_queries: string[];
  raw_text_preview?: string;
}

export interface HistoryEntry {
  date: string;
  profile: string;
  total: number;
  filepath?: string;
}

export interface HistoryResponse {
  history: HistoryEntry[];
}

export interface OutputFile {
  name: string;
  size: number;
  modified: string;
  download_url: string;
}

export interface OutputFilesResponse {
  files: OutputFile[];
}

export interface AppConfig {
  max_results_per_query: number;
  request_delay: number;
  email_enabled?: boolean;
  email_from?: string;
  email_to?: string;
  smtp_host?: string;
  smtp_port?: number;
  smtp_user?: string;
  smtp_password?: string;
}

export type ParseResumeBody = {
  file?: Blob;
};